Add test for video and playlist add rel=0
Note: We are not using Brain Monkey at this time.

Ultimately, we mocked the WP functions ourselves instead of using Brain
Monkey because as soon as the plugin is loaded, the functions fired.
Which caused a problem not being able to mock the functions early enough
with Brain Monkey.

I suspect there is a better solution to this problem, but I've not yet
found it.

At this point, Brain Monkey is unused and could be removed.

See #1